ai-service/models/regression_model.py
"""
Regression Model for demand prediction and pricing optimization
"""

import numpy as np
import pandas as pd
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, r2_score
import joblib
import os
import logging

logger = logging.getLogger(__name__)

class RegressionPredictor:
    """Linear regression model for demand prediction and pricing optimization"""
    
    def __init__(self):
        self.model = LinearRegression()
        self.scaler = StandardScaler()
        self.is_trained = False
        self.model_path = os.path.join(os.path.dirname(__file__), '..', 'models', 'regression_model.pkl')
        self.scaler_path = os.path.join(os.path.dirname(__file__), '..', 'models', 'scaler.pkl')
        
        # Try to load existing model
        self.load_model()
    
    def prepare_features(self, data):
        """Prepare features for training/prediction"""
        features = []
        
        # Extract relevant features
        if 'season' in data.columns:
            features.append(data['season'].astype('category').cat.codes)
        if 'day_of_week' in data.columns:
            features.append(data['day_of_week'])
        if 'price' in data.columns:
            features.append(data['price'])
        if 'availability' in data.columns:
            features.append(data['availability'])
        
        # Add time-based features
        if 'date' in data.columns:
            data['date'] = pd.to_datetime(data['date'])
            features.append(data['date'].dt.dayofyear)
            features.append(data['date'].dt.weekday)
        
        # Combine features
        if features:
            X = np.column_stack(features)
        else:
            # Default features if none available
            X = np.random.rand(len(data), 4)  # Placeholder
        
        return X
    
    def train(self, training_data):
        """Train the regression model"""
        try:
            # Prepare training data
            X = self.prepare_features(training_data)
            y = training_data['demand'].values if 'demand' in training_data.columns else np.random.rand(len(training_data))
            
            # Split data
            X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
            
            # Scale features
            X_train_scaled = self.scaler.fit_transform(X_train)
            X_test_scaled = self.scaler.transform(X_test)
            
            # Train model
            self.model.fit(X_train_scaled, y_train)
            
            # Evaluate
            y_pred = self.model.predict(X_test_scaled)
            mse = mean_squared_error(y_test, y_pred)
            r2 = r2_score(y_test, y_pred)
            
            logger.info(f"Model trained successfully. MSE: {mse:.4f}, R²: {r2:.4f}")
            
            # Save model
            self.save_model()
            self.is_trained = True
            
            return {
                'mse': mse,
                'r2_score': r2,
                'training_samples': len(X_train),
                'test_samples': len(X_test)
            }
            
        except Exception as e:
            logger.error(f"Error training regression model: {e}")
            raise
    
    def predict_demand(self, equipment_id, days_ahead=7):
        """Predict demand for specific equipment"""
        try:
            if not self.is_trained:
                # Return mock prediction if model not trained
                base_demand = np.random.randint(5, 20)
                return {
                    'predicted_demand': base_demand,
                    'confidence_interval': [base_demand - 2, base_demand + 2],
                    'equipment_id': equipment_id,
                    'days_ahead': days_ahead
                }
            
            # Create mock features for prediction
            features = np.array([[1, 3, 50.0, 1, 150, 2]])  # Example features
            features_scaled = self.scaler.transform(features)
            
            prediction = self.model.predict(features_scaled)[0]
            
            return {
                'predicted_demand': max(0, prediction),
                'confidence_interval': [max(0, prediction - 2), prediction + 2],
                'equipment_id': equipment_id,
                'days_ahead': days_ahead
            }
            
        except Exception as e:
            logger.error(f"Error predicting demand: {e}")
            # Return fallback prediction
            return {
                'predicted_demand': np.random.randint(5, 20),
                'confidence_interval': [3, 25],
                'equipment_id': equipment_id,
                'days_ahead': days_ahead
            }
    
    def optimize_pricing(self, equipment_id, current_price):
        """Optimize pricing based on demand elasticity"""
        try:
            # Simple price optimization based on demand elasticity
            # This is a simplified model - in reality, you'd use more sophisticated methods
            
            base_demand = 10  # Base demand at current price
            price_elasticity = -0.5  # Price elasticity coefficient
            
            # Calculate optimal price using simple elasticity model
            optimal_price = current_price * (1 + (price_elasticity * 0.1))
            
            # Ensure price is within reasonable bounds
            optimal_price = max(current_price * 0.8, min(optimal_price, current_price * 1.2))
            
            return round(optimal_price, 2)
            
        except Exception as e:
            logger.error(f"Error optimizing pricing: {e}")
            return current_price
    
    def save_model(self):
        """Save trained model and scaler"""
        try:
            os.makedirs(os.path.dirname(self.model_path), exist_ok=True)
            joblib.dump(self.model, self.model_path)
            joblib.dump(self.scaler, self.scaler_path)
            logger.info("Model and scaler saved successfully")
        except Exception as e:
            logger.error(f"Error saving model: {e}")
    
    def load_model(self):
        """Load pre-trained model and scaler"""
        try:
            if os.path.exists(self.model_path) and os.path.exists(self.scaler_path):
                self.model = joblib.load(self.model_path)
                self.scaler = joblib.load(self.scaler_path)
                self.is_trained = True
                logger.info("Model and scaler loaded successfully")
        except Exception as e:
            logger.warning(f"Could not load existing model: {e}")
            self.is_trained = False
