ai-service/requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Machine Learning and AI
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Database connectivity
PyMySQL==1.1.0
SQLAlchemy==2.0.21

# Data processing
Pillow==10.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
